move allocation dispatcher impls to ssm-mem.c
